Fix clipping error on some RTL text
There was a false assumption that the bidi level of text is only
used for visualizing the text, not for calculating its bounding
box. But the bidi level is required for shaping (indeed many
OpenType rules check for reading direction) and the glyphs used
to represent a given text may be different based on its
directionality. The effect would be that the bounding rect we
calculated for text would sometimes be too small for RTL text, and
we would end up clipping pixels.

[ChangeLog][QtGui][Text] Fixed clipping errors and too small
bounding rects for some right-to-left text.

Task-number: QTBUG-48005
Change-Id: I6479384602c69a83686886db9423dacd97942460
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 098d87c63fb638db619808af8b86c043354f62c5)
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>